Refactor ColPali service structure and update API routes
- Updated Dockerfiles to reference the new app structure.
- Removed the old app.py file and created a new modular structure with separate files for services, models, and utilities.
- Introduced a configuration module for managing settings.
- Added logging configuration with request context support.
- Created a dedicated embedding processor service for handling query and image embeddings.
- Implemented a model service for loading and managing the ColModernVBert model.
- Updated API routes to utilize the new service architecture.
- Enhanced error handling and logging throughout the application.
- Updated docker-compose.yml to comment out unused service profiles.
