+++ colpali/Dockerfile.cpu
@@ -21,4 +21,4 @@ COPY . /app
 
 EXPOSE 7000
 
-CMD ["uvicorn", "app:app", "--host", "0.0.0.0", "--port", "7000"]
+CMD ["uvicorn", "app.main:app", "--host", "0.0.0.0", "--port", "7000"]
