+++ colpali/Dockerfile.gpu
@@ -57,4 +57,4 @@ COPY . /app
 EXPOSE 7000
 
 # Start API
-CMD ["uvicorn", "app:app", "--host", "0.0.0.0", "--port", "7000"]
+CMD ["uvicorn", "app.main:app", "--host", "0.0.0.0", "--port", "7000"]
