--- colpali/app.py
@@ -1,491 +1,0 @@
-import inspect
-import logging
-import os
-from io import BytesIO
-from typing import Any, List, Optional, Tuple, Union, cast
-
-import torch
-from colpali_engine.models import ColModernVBert, ColModernVBertProcessor
-from fastapi import Body, FastAPI, File, HTTPException, Request, UploadFile, status
-from fastapi.responses import JSONResponse
-from middleware.request_id import RequestIDMiddleware
-from middleware.timing import TimingMiddleware
-from PIL import Image
-from pydantic import BaseModel
-from transformers.utils.import_utils import is_flash_attn_2_available
-
-
-# Setup logging with request ID support
-class RequestContextFilter(logging.Filter):
-    """Add request context to log records."""
-
-    def filter(self, record: logging.LogRecord) -> bool:
-        """Add request_id to record if not present."""
-        if not hasattr(record, "request_id"):
-            record.request_id = "-"
-        return True
-
-
-logging.basicConfig(
-    level=logging.INFO,
-    format="%(asctime)s | %(request_id)s | %(levelname)-8s | %(name)s | %(message)s",
-    datefmt="%Y-%m-%d %H:%M:%S",
-)
-
-# Add request context filter to root logger
-for handler in logging.root.handlers:
-    handler.addFilter(RequestContextFilter())
-
-logger = logging.getLogger(__name__)
-
-# Configuration for CPU parallelism
-CPU_THREADS = int(
-    os.getenv("CPU_THREADS", "4")
-)  # Number of torch threads for CPU inference
-ENABLE_CPU_MULTIPROCESSING = (
-    os.getenv("ENABLE_CPU_MULTIPROCESSING", "false").lower() == "true"
-)
-
-# Hugging Face model identifier (overridable via environment variable)
-MODEL_ID = os.getenv("COLPALI_MODEL_ID", "ModernVBERT/colmodernvbert-merged")
-API_VERSION = os.getenv("COLPALI_API_VERSION", "0.0.2")
-
-# Initialize FastAPI app
-app = FastAPI(
-    title="ColModernVBert Embedding API",
-    description="API for generating embeddings from images and queries",
-)
-
-
-# Global exception handler for structured error logging
-@app.exception_handler(Exception)
-async def global_exception_handler(request: Request, exc: Exception):
-    """Log unhandled exceptions with structured context."""
-    logger.error(
-        "Unhandled exception occurred",
-        exc_info=exc,
-        extra={
-            "method": request.method,
-            "path": str(request.url.path),
-            "query_params": dict(request.query_params),
-            "client_host": request.client.host if request.client else None,
-        },
-    )
-    return JSONResponse(
-        status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-        content={"detail": "Internal server error"},
-    )
-
-
-# Add middleware (order matters - first added = outermost layer)
-app.add_middleware(RequestIDMiddleware)
-app.add_middleware(TimingMiddleware)
-
-# Determine device
-device = (
-    "cuda:0"
-    if torch.cuda.is_available()
-    else "mps" if torch.backends.mps.is_available() else "cpu"
-)
-
-# Configure CPU threading for better performance
-if device == "cpu":
-    torch.set_num_threads(CPU_THREADS)
-    torch.set_num_interop_threads(CPU_THREADS)
-    logger.info(f"CPU mode: Set torch threads to {CPU_THREADS}")
-
-TORCH_DTYPE = torch.bfloat16 if device != "cpu" else torch.float32
-
-# Load model and processor
-model = cast(
-    Any,
-    ColModernVBert.from_pretrained(
-        MODEL_ID,
-        torch_dtype=TORCH_DTYPE,
-        device_map=device,
-        attn_implementation=(
-            "flash_attention_2" if is_flash_attn_2_available() else None
-        ),
-        trust_remote_code=True,
-    ).eval(),
-)
-
-_processor_loaded: Union[
-    ColModernVBertProcessor, Tuple[ColModernVBertProcessor, dict[str, Any]]
-] = ColModernVBertProcessor.from_pretrained(MODEL_ID, trust_remote_code=True)
-if isinstance(_processor_loaded, tuple):
-    processor = cast(Any, _processor_loaded[0])
-else:
-    processor = cast(Any, _processor_loaded)
-
-logger.info(f"ColModernVBert model loaded on device: {device}")
-logger.info(f"Model id: {MODEL_ID}")
-logger.info(f"Model dtype: {model.dtype}")
-logger.info(
-    f"Flash Attention 2: {'enabled' if is_flash_attn_2_available() else 'disabled'}"
-)
-
-
-def _resolve_image_token_id() -> int:
-    """Best-effort resolution of the image token id for the current processor."""
-    if hasattr(processor, "image_token_id"):
-        return int(processor.image_token_id)  # type: ignore[arg-type]
-
-    image_token = getattr(processor, "image_token", None)
-    if image_token is not None and hasattr(processor, "tokenizer"):
-        token_id = processor.tokenizer.convert_tokens_to_ids(image_token)  # type: ignore[attr-defined]
-        if token_id is not None:
-            return int(token_id)
-
-    raise AttributeError("Processor does not expose an image_token_id.")
-
-
-IMAGE_TOKEN_ID = _resolve_image_token_id()
-
-
-class QueryRequest(BaseModel):
-    queries: Union[str, List[str]]
-
-
-class QueryEmbeddingResponse(BaseModel):
-    embeddings: List[List[List[float]]]
-
-
-class Dimension(BaseModel):
-    width: int
-    height: int
-
-
-class PatchResult(BaseModel):
-    width: int
-    height: int
-    n_patches_x: Optional[int] = None
-    n_patches_y: Optional[int] = None
-    error: Optional[str] = None
-
-
-class PatchRequest(BaseModel):
-    dimensions: List[Dimension]
-
-
-class PatchBatchResponse(BaseModel):
-    results: List[PatchResult]
-
-
-class ImageEmbeddingItem(BaseModel):
-    # A single image's embeddings and the image-token boundaries
-    embedding: List[List[float]]  # [sequence_length, hidden_dim]
-    image_patch_start: int  # index where image tokens begin
-    image_patch_len: int  # number of image tokens (should equal x_patches * y_patches)
-    image_patch_indices: List[int]  # explicit positions of every image token
-
-
-class ImageEmbeddingBatchResponse(BaseModel):
-    embeddings: List[ImageEmbeddingItem]
-
-
-def load_image_from_bytes(image_bytes: bytes) -> Image.Image:
-    """Load PIL Image from bytes"""
-    return Image.open(BytesIO(image_bytes)).convert("RGB")
-
-
-def generate_query_embeddings(queries: List[str]) -> List[torch.Tensor]:
-    """Generate embeddings for text queries"""
-    device = model.device
-    with torch.no_grad():
-        batch_query = processor.process_queries(queries).to(device)
-        query_embeddings = cast(torch.Tensor, model(**batch_query))  # [batch, seq, dim]
-        # Unbind into per-sample tensors on CPU
-        return list(torch.unbind(query_embeddings.to("cpu")))
-
-
-def generate_image_embeddings_with_boundaries(
-    images: List[Image.Image],
-) -> List[ImageEmbeddingItem]:
-    """Generate embeddings for images and expose image-token boundaries."""
-    device = model.device
-    with torch.no_grad():
-        # Tokenize / encode images
-        batch_images = processor.process_images(images).to(device)
-
-        # Forward pass
-        image_embeddings = cast(
-            torch.Tensor, model(**batch_images)
-        )  # [batch, seq, dim]
-        image_embeddings = image_embeddings.to("cpu")
-
-        # Expect token ids to be present, so we can find image-token spans
-        if "input_ids" not in batch_images:
-            raise RuntimeError(
-                "Tokenizer output missing 'input_ids'; cannot compute image token boundaries."
-            )
-
-        input_ids = batch_images["input_ids"].to("cpu")  # [batch, seq]
-        image_token_id = IMAGE_TOKEN_ID
-
-        batch_items: List[ImageEmbeddingItem] = []
-        batch_size = input_ids.shape[0]
-
-        for i in range(batch_size):
-            ids = input_ids[i]  # [seq]
-            emb = image_embeddings[i]  # [seq, dim]
-
-            mask = ids.eq(image_token_id)  # bool mask for image tokens
-            indices = torch.nonzero(mask, as_tuple=True)[0]  # [num_image_tokens] or []
-            indices_list = indices.view(-1).tolist() if indices.numel() > 0 else []
-
-            if not indices_list:
-                # No image tokens found; return sentinel values
-                start = -1
-                length = 0
-            else:
-                start = int(indices_list[0])
-                length = len(indices_list)
-
-            batch_items.append(
-                ImageEmbeddingItem(
-                    embedding=emb.tolist(),
-                    image_patch_start=start,
-                    image_patch_len=length,
-                    image_patch_indices=[int(idx) for idx in indices_list],
-                )
-            )
-
-        return batch_items
-
-
-# API Endpoints
-
-
-@app.get("/")
-async def root():
-    return {"message": "ColModernVBert Embedding API", "version": API_VERSION}
-
-
-@app.get("/health")
-async def health_check():
-    """Health check endpoint"""
-    return {"status": "healthy", "device": str(model.device)}
-
-
-@app.post("/restart")
-async def restart_service():
-    """
-    Restart the service by exiting the process.
-    The container will automatically restart if configured with restart policy.
-    This is useful for stopping any ongoing processing and resetting service state.
-    """
-    import threading
-    import os
-
-    logger.info("Restart requested - initiating service shutdown")
-
-    def force_exit():
-        """Force exit in a separate thread to bypass event loop blocking"""
-        import time
-
-        time.sleep(0.1)  # Small delay to allow response to be sent
-        logger.info("Exiting process for restart")
-        os._exit(1)  # Hard exit with code 1 - terminates immediately
-
-    # Use a daemon thread to force exit regardless of event loop state
-    exit_thread = threading.Thread(target=force_exit, daemon=True)
-    exit_thread.start()
-
-    return {"status": "restarting", "message": "Service will restart momentarily"}
-
-
-@app.get("/info")
-async def version():
-    """Version endpoint"""
-    spatial_merge_size = getattr(model, "spatial_merge_size", None)
-    image_seq_len = getattr(processor, "image_seq_len", None)
-    return {
-        "version": API_VERSION,
-        "model_id": MODEL_ID,
-        "device": str(model.device),
-        "dtype": str(model.dtype),
-        "flash_attn": is_flash_attn_2_available(),
-        "spatial_merge_size": spatial_merge_size,
-        "dim": getattr(model, "dim", None),
-        "image_token_id": IMAGE_TOKEN_ID,
-        "image_seq_len": image_seq_len,
-    }
-
-
-@app.post(
-    "/patches",
-    response_model=PatchBatchResponse,
-    response_model_exclude_none=True,
-)
-async def get_n_patches(
-    request: PatchRequest = Body(
-        ..., example={"dimensions": [{"width": 1024, "height": 768}]}
-    )
-):
-    """Calculate number of patches for given image dimensions and spatial merge size
-
-    Args:
-        request: PatchRequest containing a list of dimensions with 'width' and 'height' keys
-    """
-    try:
-        results = []
-        get_n_patches_fn = getattr(processor, "get_n_patches", None)
-        supports_patches = callable(get_n_patches_fn)
-        call_kwargs: dict[str, Any] = {}
-        unsupported_reason: Optional[str] = None
-
-        if supports_patches:
-            try:
-                if get_n_patches_fn is not None:
-                    signature = inspect.signature(get_n_patches_fn)
-                else:
-                    signature = None
-            except (TypeError, ValueError):
-                signature = None
-
-            if signature is not None:
-                for name, param in signature.parameters.items():
-                    if name == "image_size":
-                        continue
-
-                    value: Any = None
-                    if name in {"patch_size", "spatial_merge_size"}:
-                        value = getattr(model, "spatial_merge_size", None)
-                    elif hasattr(processor, name):
-                        value = getattr(processor, name)
-                    elif hasattr(model, name):
-                        value = getattr(model, name)
-
-                    if value is None:
-                        if param.default is inspect._empty:
-                            supports_patches = False
-                            unsupported_reason = f"Required parameter '{name}' is not provided by the current model."
-                            break
-                        continue
-
-                    call_kwargs[name] = value
-
-            if supports_patches and signature is None:
-                # Unable to inspect signature safely; proceed without extra kwargs.
-                call_kwargs = {}
-
-        if not supports_patches:
-            unsupported_reason = (
-                unsupported_reason
-                or "Patch estimation is not available for the current model."
-            )
-
-        for dim in request.dimensions:
-            try:
-                if not supports_patches:
-                    raise NotImplementedError(unsupported_reason)
-
-                image_size = (dim.width, dim.height)
-                try:
-                    n_patches_x, n_patches_y = get_n_patches_fn(  # type: ignore[misc]
-                        image_size, **call_kwargs
-                    )
-                except NotImplementedError:
-                    raise
-
-                results.append(
-                    PatchResult(
-                        width=dim.width,
-                        height=dim.height,
-                        n_patches_x=int(n_patches_x),
-                        n_patches_y=int(n_patches_y),
-                    )
-                )
-            except Exception as e:
-                results.append(
-                    PatchResult(
-                        width=dim.width,
-                        height=dim.height,
-                        error=str(e),
-                    )
-                )
-        return {"results": results}
-    except Exception as e:
-        raise HTTPException(
-            status_code=500, detail=f"Error processing patch request: {str(e)}"
-        )
-
-
-@app.post("/embed/queries", response_model=QueryEmbeddingResponse)
-async def embed_queries(request: QueryRequest):
-    """Generate embeddings for text queries"""
-    try:
-        queries = (
-            [request.queries] if isinstance(request.queries, str) else request.queries
-        )
-        if not queries:
-            raise HTTPException(status_code=400, detail="No queries provided")
-
-        # Run embedding generation in thread pool to avoid blocking event loop
-        import asyncio
-        from concurrent.futures import ThreadPoolExecutor
-
-        # Use a shared thread pool executor for GPU operations
-        if not hasattr(embed_queries, '_executor'):
-            embed_queries._executor = ThreadPoolExecutor(max_workers=1, thread_name_prefix="query")
-
-        embeddings_tensors = await asyncio.get_event_loop().run_in_executor(
-            embed_queries._executor,
-            generate_query_embeddings,
-            queries
-        )
-        embeddings_list = [embedding.tolist() for embedding in embeddings_tensors]
-        return QueryEmbeddingResponse(embeddings=embeddings_list)
-
-    except Exception as e:
-        raise HTTPException(
-            status_code=500, detail=f"Error generating query embeddings: {str(e)}"
-        )
-
-
-@app.post("/embed/images", response_model=ImageEmbeddingBatchResponse)
-async def embed_images(files: List[UploadFile] = File(...)):
-    """Generate embeddings for uploaded images + image-token boundaries."""
-    try:
-        if not files:
-            raise HTTPException(status_code=400, detail="No images provided")
-
-        images: List[Image.Image] = []
-        for file in files:
-            content_type = file.content_type or ""
-            if not content_type.startswith("image/"):
-                raise HTTPException(
-                    status_code=400, detail=f"File {file.filename} is not an image"
-                )
-            image_bytes = await file.read()
-            images.append(load_image_from_bytes(image_bytes))
-
-        # Run embedding generation in thread pool to avoid blocking event loop
-        # This allows /restart endpoint to be processed immediately during cancellation
-        import asyncio
-        from concurrent.futures import ThreadPoolExecutor
-
-        # Use a shared thread pool executor for GPU operations
-        if not hasattr(embed_images, '_executor'):
-            embed_images._executor = ThreadPoolExecutor(max_workers=1, thread_name_prefix="embed")
-
-        items = await asyncio.get_event_loop().run_in_executor(
-            embed_images._executor,
-            generate_image_embeddings_with_boundaries,
-            images
-        )
-        return ImageEmbeddingBatchResponse(embeddings=items)
-
-    except HTTPException:
-        raise
-    except Exception as e:
-        raise HTTPException(
-            status_code=500, detail=f"Error generating image embeddings: {str(e)}"
-        )
-
-
-if __name__ == "__main__":
-    import uvicorn
-
-    uvicorn.run("app:app", host="0.0.0.0", port=7000, reload=True, workers=1)
