+++ colpali/main.py
@@ -1,0 +1,8 @@
+"""Entry point for ColPali service."""
+
+import uvicorn
+
+from app.main import app
+
+if __name__ == "__main__":
+    uvicorn.run(app, host="0.0.0.0", port=7000, workers=1)
